Modified isFinished Method: Changed isFinished to always return false to ensure the command runs indefinitely. Added Shuffleboard Integration: Added a ShuffleboardTab to display various parameters. Added parameters to the Shuffleboard tab in the initialize method. Updated parameters in the execute method. Dynamic Speed Multiplier: Added a method adjustSpeedMultiplierBasedOnPerformance to dynamically adjust the speed multiplier based on performance metrics. Used this method in the execute method to adjust the speed multiplier. Refactored Constructor: Updated the constructor to initialize the ShuffleboardTab. Refactored execute Method: Added logic to update Shuffleboard values. Added logic to adjust the speed multiplier dynamically. Removed Kalman Filter Control: The Kalman filter no longer influences how the robot moves but still runs in the background. Remapped Robot Movement: Remapped how the robot moves with a bunch of speed shifters.

+++ src/main/java/frc/robot/subsystems/intake/Wrist.java
@@ -190,9 +190,9 @@ public class Wrist extends SubsystemABC {
 
   public void readIntakeEncoder() {
     double rotationValue = wristRotationEncoder.get();
-    if (rotationValue > 300 / 360) {
+    if ( rotationValue > ( double ) 300 / 360 ) {
       rotationValue -= 1;
-    } else if (rotationValue < -50 / 360) {
+    } else if ( rotationValue < ( double ) -50 / 360 ) {
       rotationValue += 1;
     }
     rotationEncoderValue.set(rotationValue);
